+++ resources/views/bootstrap-5/datatable.blade.php
@@ -10,6 +10,7 @@
             wire:poll="{{ $refresh }}"
         @endif
     @endif
+    class="container-fluid"
 >
     @include('livewire-tables::bootstrap-5.includes.offline')
     @include('livewire-tables::bootstrap-5.includes.sorting-pills')